# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Tarefas",
        "version": "1.0"
    },
    "paths": {
        "/api/tasks": {
            "get": {
                "summary": "Listar todas as tarefas",
                "operationId": "471d938a3bfaeb6fbd981d7c3e2728b0",
                "responses": {
                    "200": {
                        "description": "Lista de tarefas"
                    }
                }
            },
            "post": {
                "summary": "Criar uma nova tarefa",
                "operationId": "f803ebdb2300acdb7e93bbe8d0b9a2e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Nova Tarefa"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição da nova tarefa"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": false
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tarefa criada"
                    }
                }
            }
        },
        "/api/tasks/{id}": {
            "get": {
                "summary": "Obter detalhes de uma tarefa",
                "operationId": "90b3cd4d2d4b416dc989d2055dbed06c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes da tarefa"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                }
            },
            "put": {
                "summary": "Atualizar uma tarefa",
                "operationId": "1aaf37e74360ef4c10fadaee02ae77fb",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Tarefa Atualizada"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição atualizada"
                                    },
                                    "completed": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tarefa atualizada"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                }
            },
            "delete": {
                "summary": "Deletar uma tarefa",
                "operationId": "ad582480cd57c2a4e3b6a426943eeca5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tarefa deletada"
                    },
                    "404": {
                        "description": "Tarefa não encontrada"
                    }
                }
            }
        }
    }
}